+++ releases.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Unreleased
+
+  - Introduce `bake utopia:project:agent:context:update` command to update the agent context from the guides in the project.
+
 ## v0.33.2
 
   - Fixed handling of segmented code guides when rendered into a `readme.md` file.